load-bigquery/Kylie/requirements.txt
pandas
numpy
sklearn
joblib
coloredlogs
google-cloud-storage
google-cloud-bigquery
openpyxl
